moved file status functionality into Git::Repository without using a separate class
