+++ libGitWrap/RepositoryPrivate.hpp
@@ -44,7 +44,6 @@ namespace Git
         public:
             git_repository* mRepo;
             IndexPrivate*   mIndex;
-            StatusPrivate*  mStatus;
         };
 
     }
